Hello PICAFlow users!
This time, a very useful update :)

This version now features a novel Shiny application tool which allows to interactively visualize and edit the peak values for each sample and parameter. It will be much better than opening the tabular text files in Excel and editing them by hand to correct for misidentification of peaks. This will ensure a better and more intuitive workflow.

This functionnality is totally optinal though and can be entirely skipped if desired.

Please consult the tutorial to have all the details about its usage.